requirements.txt
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
clip==1.0
cryptography==39.0.1
einops==0.6.1
faiss-gpu==1.7.2
filelock==3.9.0
fsspec==2023.5.0
ftfy==6.1.1
gmpy2==2.1.2
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kornia==0.6.12
MarkupSafe==2.1.1
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
networkx==2.8.4
numpy==1.24.3
packaging==23.1
Pillow==9.4.0
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
PyYAML==6.0
regex==2023.5.5
requests==2.29.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sympy==1.11.1
threadpoolctl==3.1.0
timm==0.9.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
